requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-cors==0.8.1
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
astor==0.8.1
attrs==25.3.0
bitsandbytes==0.45.4
blake3==1.0.4
blessed==1.20.0
boto3==1.37.27
botocore==1.37.27
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.6
compressed-tensors==0.9.2
contourpy==1.3.1
cupy-cuda12x==13.4.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.4.1
deepspeed==0.16.5
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
dotmap==1.3.30
einops==0.8.1
elastic-transport==8.17.1
elasticsearch==8.17.2
email_validator==2.2.0
fancycompleter==0.9.1
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.2.1
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
gpustat==1.1.1
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
hf-xet==1.0.3
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
json5==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.3.18
lark==1.2.2
llguidance==0.7.13
llvmlite==0.44.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
mistral_common==1.5.4
mpi4py==3.1.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nanobind==2.6.1
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.2
omegaconf==2.3.0
openai==1.68.2
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
orjson==3.10.15
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
pdbpp==0.10.3
peft==0.15.0
pillow==11.1.0
pip==25.0
platformdirs==4.3.7
pluggy==1.5.0
prometheus_client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.8.1
Pygments==2.19.1
pylatexenc==2.10
pynvml==12.0.0
pyparsing==3.2.3
pyrepl==0.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.4.0
ragas==0.2.14
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.1
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.25.1
setproctitle==1.3.5
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
swankit==0.1.7
swanlab==0.5.5
sympy==1.13.1
tenacity==9.0.0
tensordict==0.6.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchdata==0.11.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.1
transformers-stream-generator==0.0.5
triton==3.2.0
trl==0.16.0
typer==0.15.2
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
verl==0.3.0.post1
virtualenv==20.30.0
vllm==0.8.3
wandb==0.19.9
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
wmctrl==0.5
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.17
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
